+++ testutil/daemon/daemon_unix.go
@@ -11,9 +11,19 @@ import (
 	"syscall"
 	"testing"
 
+	"github.com/docker/docker/pkg/mount"
 	"golang.org/x/sys/unix"
 	"gotest.tools/v3/assert"
 )
+
+// cleanupMount unmounts the daemon root directory, or logs a message if
+// unmounting failed.
+func cleanupMount(t testing.TB, d *Daemon) {
+	t.Helper()
+	if err := mount.Unmount(d.Root); err != nil {
+		d.log.Logf("[%s] unable to unmount daemon root (%s): %v", d.id, d.Root, err)
+	}
+}
 
 func cleanupNetworkNamespace(t testing.TB, d *Daemon) {
 	t.Helper()
